Move pagination functions to their own module
This just moves pagination related functions apply_limit and
apply_order_by to their own module so they can be imported and used from
other parts of the application.

The only code changes are prefixing "pagination." in front of the called
functions. The functions themselves have just been moved without any
modifications.
